# Conflicts:
#	ibizpms-core/src/main/java/cn/ibizlab/pms/core/util/ibizzentao/helper/TaskHelper.java

+++ app_Web/src/widgets/todo/main-form/main-form-model.ts
@@ -73,7 +73,7 @@ export default class MainModel {
       {
         name: 'idvalue',
         prop: 'idvalue',
-        dataType: 'INT',
+        dataType: 'BIGINT',
       },
       {
         name: 'formitem3',
